# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Talis - Celestia's infrastructure management service",
        "title": "Talis API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/instances/{instance_id}/tasks": {
            "get": {
                "description": "Returns a list of tasks for a specific instance with optional filtering and pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "List tasks for an instance",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Instance ID",
                        "name": "instance_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Filter by task action",
                        "name": "action",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items to return (default 10)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items to skip (default 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of tasks",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_celestiaorg_talis_internal_types.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_celestiaorg_talis_internal_types.TaskListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/github_com_celestiaorg_talis_internal_types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_celestiaorg_talis_internal_types.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_celestiaorg_talis_internal_types.ErrorResponse": {
            "type": "object",
            "properties": {
                "details": {},
                "error": {
                    "type": "string"
                }
            }
        },
...
        "github_com_celestiaorg_talis_internal_types.PaginationResponse": {
            "type": "object",
            "properties": {
                "limit": {
                    "description": "Number of items per page",
                    "type": "integer"
                },
                "offset": {
                    "description": "Offset from start of results",
                    "type": "integer"
                },
                "page": {
                    "description": "Current page number",
                    "type": "integer"
                },
                "total": {
                    "description": "Total number of values",
                    "type": "integer"
                }
            }
        },
...
        "github_com_celestiaorg_talis_internal_types.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {}
            }
        },
        "github_com_celestiaorg_talis_internal_types.TaskListResponse": {
            "type": "object",
            "properties": {
                "pagination": {
                    "$ref": "#/definitions/github_com_celestiaorg_talis_internal_types.PaginationResponse"
                },
                "rows": {
                    "type": "array",
                    "items": {}
                }
            }
        },
...
    }
}